(#1218) Remove 2d gridscan test and dedup fixtures, remove obsolete TODOs as per PR feedback
