+++ src/hyperion/external_interaction/nexus/nexus_utils.py
@@ -133,7 +133,6 @@ def create_beam_and_attenuator_parameters(
         tuple[Beam, Attenuator]: Descriptions of the beam and attenuator for nexgen.
     """
     return (
-        # TODO 1173 Get this data from events rather than ispyb_params
         Beam(convert_eV_to_angstrom(energy_kev * 1000), flux),  # pyright: ignore
         Attenuator(transmission_fraction),  # pyright: ignore
     )
